refactor: recoil state 정리
- public 삭제
- initialState 파일 추가
